fix: try splitting out types from default export
