+++ src/index.ts
@@ -1,3 +1,4 @@
+import type { FetchHAROptions, RequestInitWithDuplex } from './types.js';
 import type { DataURL as npmDataURL } from '@readme/data-urls';
 import type { Har } from 'har-format';
 
@@ -22,25 +23,6 @@ if (!globalThis.File) {
   } catch (e) {
     throw new Error('The File API is required for this library. https://developer.mozilla.org/en-US/docs/Web/API/File');
   }
-}
-
-interface RequestInitWithDuplex extends RequestInit {
-  /**
-   * `RequestInit#duplex` does not yet exist in the TS `lib.dom.d.ts` definition yet the native
-   * fetch implementation in Node 18+, `undici`, requires it for certain POST payloads.
-   *
-   * @see {@link https://github.com/microsoft/TypeScript-DOM-lib-generator/issues/1483}
-   * @see {@link https://github.com/nodejs/node/issues/46221}
-   * @see {@link https://fetch.spec.whatwg.org/#request-class}
-   * @see {@link https://github.com/microsoft/TypeScript/blob/main/lib/lib.dom.d.ts}
-   */
-  duplex?: 'half';
-}
-
-export interface FetchHAROptions {
-  files?: Record<string, Blob | Buffer>;
-  init?: RequestInitWithDuplex;
-  userAgent?: string;
 }
 
 type DataURL = npmDataURL & {
